continue switching rest of controllers to SOA + little update in getAllNotificationForUser method
