+++ app/Services/NotificationService.php
@@ -3,13 +3,19 @@
 namespace App\Services;
 
 use Illuminate\Notifications\DatabaseNotification;
-use Illuminate\Support\Facades\Auth;
 
 class NotificationService
 {
     public function getAllNotificationsForUser()
     {
-        return Auth::user()->notifications;
+        $notifications = auth()->user()->notifications;
+
+        return $notifications->map(function($notification) {
+            return [
+                'id' => $notification->id,
+                'title' => $notification->data['title'] ?? 'No Title',
+            ];
+        });
     }
 
     public function markAsRead($id)
